make base class abstract, move two checking functions
