+++ sktime/classification/tests/test_base.py
@@ -126,30 +126,30 @@ def test_check_capabilities():
     """
     handles_none = _DummyClassifier()
 
-    handles_none.check_capabilities(False, False, False)
+    handles_none._check_capabilities(False, False, False)
     with pytest.raises(ValueError, match=missing_message):
-        handles_none.check_capabilities(True, True, True)
-        handles_none.check_capabilities(True, True, False)
-        handles_none.check_capabilities(True, False, False)
-        handles_none.check_capabilities(True, False, True)
+        handles_none._check_capabilities(True, True, True)
+        handles_none._check_capabilities(True, True, False)
+        handles_none._check_capabilities(True, False, False)
+        handles_none._check_capabilities(True, False, True)
     with pytest.raises(ValueError, match=multivariate_message):
-        handles_none.check_capabilities(False, True, True)
-        handles_none.check_capabilities(False, True, False)
-        handles_none.check_capabilities(False, False, True)
+        handles_none._check_capabilities(False, True, True)
+        handles_none._check_capabilities(False, True, False)
+        handles_none._check_capabilities(False, False, True)
     with pytest.raises(ValueError, match=unequal_message):
-        handles_none.check_capabilities(False, False, True)
+        handles_none._check_capabilities(False, False, True)
 
     handles_all = _DummyHandlesAllInput()
-    handles_all.check_capabilities(False, False, False)
-    handles_all.check_capabilities(False, False, False)
-    handles_all.check_capabilities(True, True, True)
-    handles_all.check_capabilities(True, True, False)
-    handles_all.check_capabilities(True, False, True)
-    handles_all.check_capabilities(False, True, True)
-    handles_all.check_capabilities(True, False, False)
-    handles_all.check_capabilities(False, True, False)
-    handles_all.check_capabilities(False, False, True)
-    handles_all.check_capabilities(False, False, False)
+    handles_all._check_capabilities(False, False, False)
+    handles_all._check_capabilities(False, False, False)
+    handles_all._check_capabilities(True, True, True)
+    handles_all._check_capabilities(True, True, False)
+    handles_all._check_capabilities(True, False, True)
+    handles_all._check_capabilities(False, True, True)
+    handles_all._check_capabilities(True, False, False)
+    handles_all._check_capabilities(False, True, False)
+    handles_all._check_capabilities(False, False, True)
+    handles_all._check_capabilities(False, False, False)
 
 
 def test_convert_input():
@@ -168,33 +168,33 @@ def test_convert_input():
     test_X1 = np.random.uniform(-1, 1, size=(cases, length))
     test_X2 = np.random.uniform(-1, 1, size=(cases, 2, length))
     tester = _DummyClassifier()
-    tempX = tester.convert_X(test_X1)
+    tempX = tester._convert_X(test_X1)
     assert tempX.shape[0] == cases and tempX.shape[1] == 1 and tempX.shape[2] == length
-    tempX = tester.convert_X(test_X2)
+    tempX = tester._convert_X(test_X2)
     assert tempX.shape[0] == cases and tempX.shape[1] == 2 and tempX.shape[2] == length
     instance_list = []
     for _ in range(0, cases):
         instance_list.append(pd.Series(np.random.randn(10)))
     test_X3 = _create_example_dataframe(cases=cases, dimensions=1, length=length)
     test_X4 = _create_example_dataframe(cases=cases, dimensions=3, length=length)
-    tempX = tester.convert_X(test_X3)
+    tempX = tester._convert_X(test_X3)
     assert tempX.shape[0] == cases and tempX.shape[1] == 1 and tempX.shape[2] == length
-    tempX = tester.convert_X(test_X4)
+    tempX = tester._convert_X(test_X4)
     assert tempX.shape[0] == cases and tempX.shape[1] == 3 and tempX.shape[2] == length
     tester = _DummyConvertPandas()
-    tempX = tester.convert_X(test_X1)
+    tempX = tester._convert_X(test_X1)
     assert isinstance(tempX, pd.DataFrame)
     assert tempX.shape[0] == cases
     assert tempX.shape[1] == 1
-    tempX = tester.convert_X(test_X2)
+    tempX = tester._convert_X(test_X2)
     assert isinstance(tempX, pd.DataFrame)
     assert tempX.shape[0] == cases
     assert tempX.shape[1] == 2
     test_y1 = np.random.randint(0, 1, size=(cases))
     test_y2 = pd.Series(test_y1)
-    tempY = tester.convert_y(test_y1)
+    tempY = tester._convert_y(test_y1)
     assert isinstance(tempY, np.ndarray)
-    tempY = tester.convert_y(test_y2)
+    tempY = tester._convert_y(test_y2)
     assert isinstance(tempY, np.ndarray)
